Better implementation of "Touch control full travel", where, only for the onscreen touch controller, the joystick used for thrust goes all the way to the bottom.
This simply remaps the range reported to the Controls class from (-1 to 1) to (0 to 1). This is mostly to avoid complexifying many other cases that assume that resetting both analog sticks to 0 is the correct action.  Unfortunately it means other types of controllers would need to implement this themselves, if they wanted to.